backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
cerebras-cloud-sdk
chromadb
sentence-transformers
